109 added fullName/role and rename someFiles
