+++ frontend/challedu-app/src/components/AllUsersList.js
@@ -17,7 +17,7 @@ function AllUsersList() {
         console.log(err);
         setLoad(true);
       })
-
+//  eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   if (load) {
