+++ frontend/challedu-app/src/components/Ngos.js
@@ -20,6 +20,7 @@ function Ngos() {
                 console.log(err);
                 setLoad(true);
             })
+            //  eslint-disable-next-line react-hooks/exhaustive-deps
     }, []);
 
     if (load) {
