+++ frontend/challedu-app/src/components/Users.js
@@ -20,6 +20,7 @@ function Users() {
                 console.log(err);
                 setLoad(true);
             })
+            //  eslint-disable-next-line react-hooks/exhaustive-deps
     }, []);
 
     if (load) {
